Reorganize `common`
