+++ src/App.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { SocketIO } from 'boardgame.io/multiplayer';
 
 import { MatchInfo, Matchmaker, debugMatchInfo } from 'lobby';
-import { IN_PROD } from 'config';
+import { IN_PROD } from 'common/config';
 import { Machikoro } from 'game';
 import { MachikoroBoard } from 'board';
 import { version } from 'version';
