+++ src/game/establishments/main.ts
@@ -2,7 +2,7 @@
 // Utility functions for establishments.
 //
 
-import { assertUnreachable } from 'common';
+import { assertUnreachable } from 'common/typescript';
 
 import * as Meta from './metadata';
 import * as Meta2 from './metadata2';
